+++ Client/.env
@@ -1,0 +1,3 @@
+DEVICE_ID=Laptop-Nitya
+SERVER_URL=http://localhost:8000/api/devices/data
+SEND_INTERVAL=10
